+++ ergo_script/src/lib.rs
@@ -266,33 +266,6 @@ mod test {
         Ok(())
     }
 
-    /*
-    #[test]
-    fn match_expr() -> Result<(), String> {
-        script_eval_to(
-            "match [1,2,3] { {^keys} -> :keys, [a,b] -> :a, [a,b,c] -> :a }",
-            SRString("1"),
-        )?;
-        script_eval_to(
-            "match [1,2,3] { a -> :a, [a,b,c] -> :b }",
-            SRArray(&[SRString("1"), SRString("2"), SRString("3")]),
-        )?;
-        Ok(())
-    }
-
-    #[test]
-    fn match_failure() -> Result<(), String> {
-        script_result_fail("match {a=[1,2]} { {b} -> :b, [a,b] -> :b }")
-    }
-
-    #[test]
-    fn forced_match() -> Result<(), String> {
-        script_eval_to("match {a=[1,2]} { {b} -> :b, [a,b] -> :b }; ()", SRUnit)?;
-        script_fail("!match {a=[1,2]} { {b} -> :b, [a,b] -> :b }")?;
-        Ok(())
-    }
-    */
-
     #[test]
     fn if_expr() -> Result<(), String> {
         script_eval_to(":t = 1; :f = 2; if () :t :f", SRString("2"))?;